+++ library/src/modules/helpers.js
@@ -8,6 +8,15 @@ function makeRandomId() {
 	return result;
 }
 
+function toKebabCase(string) {
+	return string && string
+    .match(/[A-Z]{2,}(?=[A-Z][a-z]+[0-9]*|\b)|[A-Z]?[a-z]+[0-9]*|[A-Z]|[0-9]+/g)
+    .map(x => x.toLowerCase())
+    .join('-')
+}
+
+
 export {
-	makeRandomId
+	makeRandomId,
+	toKebabCase
 }