+++ library/src/App.vue
@@ -33,7 +33,7 @@
         //- TestButtons
         //- TestAutocomplete
         //- TestStepper
-        TestSlatTable
+        //- TestSlatTable
         //- TestSlat
         //- TestFileChooser
         //- TestCheckbox
@@ -53,6 +53,7 @@
         //- TestTabs
         //- TestPills
         //- TestTagsField
+        TestTooltip
 </template>
 
 <script>
@@ -86,6 +87,7 @@ import TestDateField from './test-components/TestDateField'
 import TestTabs from './test-components/TestTabs'
 import TestPills from './test-components/TestPills'
 import TestTagsField from './test-components/TestTagsField'
+import TestTooltip from './test-components/TestTooltip'
 import { setTheme } from './modules/app/theme/theme'
 
 export default {
@@ -120,7 +122,8 @@ export default {
     TestDateField,
     TestTabs,
     TestPills,
-    TestTagsField
+    TestTagsField,
+    TestTooltip
   },
   data: () => ({
     loggedIn: false,