+++ library/src/components/filter/FilterHeader.vue
@@ -2,6 +2,15 @@
   .nio-filter-header
     .title-description
       .filter-title.nio-h4.text-primary-darker {{ title }}
+        .filter-tooltip
+          NioTooltip(
+            v-if="tooltip"
+            :heading="tooltip.heading"
+            :message="tooltip.message"
+            :linkText="tooltip.linkText"
+            :linkHref="tooltip.linkHref"
+          )
+          slot(name="header-tooltip" v-else)
       .description.nio-p.text-primary-dark {{ description }}
     .filter-value
       .value(v-for="item in value")
@@ -18,13 +27,15 @@
 <script>
 
 import NioIcon from '../icon/Icon'
+import NioTooltip from '../Tooltip'
 
 export default {
   name: 'nio-filter-header',
   props: {
      "title": { type: String, required: true },
      "description": { type: String, required: false, default: null },
-     "value": { type: Array, required: false, default: null}
+     "value": { type: Array, required: false, default: null},
+     "tooltip": { type: Object, required: false }
   },
   data: () => ({
     
@@ -32,7 +43,7 @@ export default {
   methods: {
     
   },
-  components: { NioIcon }
+  components: { NioIcon, NioTooltip }
 }
 </script>