+++ library/src/components/icon/Icon.vue
@@ -52,6 +52,7 @@ import { faEye } from './fa-icons/eye'
 import { faEyeSlash } from './fa-icons/eye-slash'
 import { faExclamation } from './fa-icons/exclamation'
 import { faCheckCircle } from './fa-icons/check-circle'
+import { faInfoCircle } from './fa-icons/info'
  
 export default {
   name: 'nio-icon',
@@ -102,6 +103,7 @@ export default {
     library.add(faEyeSlash)
 		library.add(faExclamation)
 		library.add(faCheckCircle)
+		library.add(faInfoCircle)
     this.icon = this.iconLibrary[this.name]
     this.$emit('mounted')
   },