+++ CHANGELOG.md
@@ -48,3 +48,4 @@ Added **NioTagsField**
 v0.3.10
 Added validations to numerical unconstrained vilter variant
 Added **utility-minus** icon
+Add dynamic step-name class to NioStep