+++ CHANGELOG.md
@@ -49,3 +49,5 @@ v0.3.10
 Added validations to numerical unconstrained vilter variant
 Added **utility-minus** icon
 Add dynamic step-name class to NioStep
+Fixed currency input masking to two decimal places
+Changed binding of slider text/number field values to apply on Enter keyup