+++ CHANGELOG.md
@@ -54,3 +54,4 @@ Changed binding of slider text/number field values to apply on Enter keyup
 Added support for v-model biding on **NioExpansionPanels**
 Updated **NioStepper** to support multiword step names
 Added **ctaButtons** prop to **NioSidenav** to display CTAs for apps
+Added **headerModules** prop to **NioSlatTable** to simplify customizing table header