+++ CHANGELOG.md
@@ -55,3 +55,4 @@ Added support for v-model biding on **NioExpansionPanels**
 Updated **NioStepper** to support multiword step names
 Added **ctaButtons** prop to **NioSidenav** to display CTAs for apps
 Added **headerModules** prop to **NioSlatTable** to simplify customizing table header
+Updated primary colors