+++ CHANGELOG.md
@@ -59,3 +59,6 @@ Updated primary colors
 Added **NioTooltip** component
 Added filter tooltip optional object in config to add tooltip to **NioFilterHeader** and provided custom tooltip slot
 Added optional **fluid-width** attribute to **NioTextField**
+Bug Fix: fixed two-way binding on **NioTextField** and **NioSlider**
+
+v0.3.12